requirements.txt
streamlit
requests
python-dotenv
openai==0.28
pydantic